modified - float to double in Remove Bitcoin

+++ BitcoinMillionaire/BitcoinMillionaire/RemoveBitcoin/RemoveBitcoinViewController.swift
@@ -40,7 +40,7 @@ extension RemoveBitcoinViewController {
     
     @objc func removeBitcoin() {
         let amountText = textFieldText.value
-        guard let amount = Float(amountText) else { return }
+        guard let amount = Double(amountText) else { return }
         switch presenter.removeBitcoin(amount: Double(amount)) {
         case .success(_):
             presenter.dismiss(self, animated: true)
